requirements.txt
torch>=2.0.0
torchvision>=0.15.0
hydra-core>=1.3.0
omegaconf>=2.3.0
numpy>=1.24.0
pillow>=9.0.0
tqdm>=4.65.0
tensorboard>=2.13.0
